feat(terraform): add main.tf and provider.tf files
Add main.tf file to define an AWS S3 bucket resource for the distribution origin. The bucket name is dynamically generated based on the Terraform workspace. Also, add tags to the bucket.

Add provider.tf file to configure the AWS provider. Set the required provider as "hashicorp/aws" and configure the backend to store the Terraform state in an S3 bucket named "livebus-terraform-state-v2" with the key "terraform.tfstate". Set the AWS region to "us-east-1".